bin/functional-tests/requirements.txt
pytest==6.0.1
testinfra==5.2.2
docker==4.3.1
kubernetes==11.0.0
PyYAML==5.3.1
